backend/requirements.txt
acres==0.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
boto3==1.39.2
botocore==1.39.2
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.5
dataclasses-json==0.6.7
distro==1.9.0
etelemetry==0.3.1
Faker==37.4.0
fastapi==0.116.1
filelock==3.18.0
firebase-admin==6.9.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.73.1
gunicorn==21.2.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
http_ece==1.2.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.67
langchain-openai==0.3.27
langchain-pinecone==0.2.8
langchain-tests==0.3.20
langchain-text-splitters==0.3.8
langgraph==0.5.1
langgraph-checkpoint==2.1.0
langgraph-checkpoint-postgres==2.0.21
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.72
langsmith==0.4.4
looseversion==1.3.0
lxml==6.0.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
msgpack==1.1.1
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==2.3.1
openai==1.93.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.1
pathlib==1.0.1
pillow==11.3.0
pinecone==7.3.0
pinecone-client==6.0.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
prov==2.1.1
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
puremagic==1.30
py-cpuinfo==9.0.0
py-vapid==1.9.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.3
pyparsing==3.2.3
PyPDF2==3.0.1
pytesseract==0.3.13
pytest==8.4.1
pytest-asyncio==0.26.0
pytest-benchmark==5.1.0
pytest-codspeed==3.2.0
pytest-recording==0.13.4
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
pywebpush==2.0.3
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
redis==6.2.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.13.0
schedule==1.2.2
scipy==1.16.0
simple-websocket==1.1.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.47.2
stripe==12.3.0
syrupy==4.9.1
tenacity==9.1.2
textblob==0.19.0
tiktoken==0.9.0
tqdm==4.67.1
traits==7.0.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
email-validator==2.2.0
psycopg2-binary
pdfkit==1.0.0
markdown==3.5.2
EbookLib==0.18
weasyprint==60.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
vcrpy==7.0.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
aiohttp